requirements.txt
Jinja2==3.1.2
WeasyPrint==54.0
inflect-7.3.0
